requirements.txt
psycopg2-binary
redis
sqlalchemy
requests
pandas
PySide6
python-dotenv